Run `npm ci --dev` as root in CI
Since we're using a git url dependency with a build step, npm needs to
check out the dependency and build it while installing packages. For
some reason the runner user doesn't have access to the directory npm
wants to do this in, so it errors out. The GitHub docs say we have
access to password-less sudo (after we install sudo), so this should
work.