Use an uncaughtException event handler to prevent crashing
Ideally we would catch and handle all exceptions, but there are a few
unhandled promise rejections that still plague us. It should be fine
more often than not to just let whatever bot interaction the exception
came from fail and continue the event loop. So, for now, let's just log
and suppress the process exit.